+++ protocols/identify/tests/smoke.rs
@@ -6,10 +6,13 @@ use libp2p_swarm_test::SwarmExt;
 use std::collections::HashSet;
 use std::iter;
 use std::time::{Duration, Instant};
+use tracing_subscriber::EnvFilter;
 
 #[async_std::test]
 async fn periodic_identify() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let mut swarm1 = Swarm::new_ephemeral(|identity| {
         identify::Behaviour::new(
@@ -83,7 +86,9 @@ async fn periodic_identify() {
 }
 #[async_std::test]
 async fn only_emits_address_candidate_once_per_connection() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let mut swarm1 = Swarm::new_ephemeral(|identity| {
         identify::Behaviour::new(
@@ -153,7 +158,9 @@ async fn only_emits_address_candidate_once_per_connection() {
 
 #[async_std::test]
 async fn identify_push() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let mut swarm1 = Swarm::new_ephemeral(|identity| {
         identify::Behaviour::new(identify::Config::new("a".to_string(), identity.public()))
@@ -203,7 +210,9 @@ async fn identify_push() {
 
 #[async_std::test]
 async fn discover_peer_after_disconnect() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let mut swarm1 = Swarm::new_ephemeral(|identity| {
         identify::Behaviour::new(identify::Config::new("a".to_string(), identity.public()))
@@ -254,7 +263,9 @@ async fn discover_peer_after_disconnect() {
 
 #[async_std::test]
 async fn configured_interval_starts_after_first_identify() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let identify_interval = Duration::from_secs(5);