+++ protocols/mdns/src/behaviour.rs
@@ -282,6 +282,7 @@ where
             .on_swarm_event(&event);
     }
 
+    #[tracing::instrument(level = "trace", name = "ConnectionHandler::poll", skip(self, cx))]
     fn poll(
         &mut self,
         cx: &mut Context<'_>,
@@ -310,18 +311,20 @@ where
                             Ok(iface_state) => {
                                 e.insert(P::spawn(iface_state));
                             }
-                            Err(err) => log::error!("failed to create `InterfaceState`: {}", err),
+                            Err(err) => {
+                                tracing::error!("failed to create `InterfaceState`: {}", err)
+                            }
                         }
                     }
                 }
                 Ok(IfEvent::Down(inet)) => {
                     if let Some(handle) = self.if_tasks.remove(&inet.addr()) {
-                        log::info!("dropping instance {}", inet.addr());
+                        tracing::info!(instance=%inet.addr(), "dropping instance");
 
                         handle.abort();
                     }
                 }
-                Err(err) => log::error!("if watch returned an error: {}", err),
+                Err(err) => tracing::error!("if watch returned an error: {}", err),
             }
         }
         // Emit discovered event.
@@ -337,7 +340,7 @@ where
             {
                 *cur_expires = cmp::max(*cur_expires, expiration);
             } else {
-                log::info!("discovered: {} {}", peer, addr);
+                tracing::info!(%peer, address=%addr, "discovered peer on address");
                 self.discovered_nodes.push((peer, addr.clone(), expiration));
                 discovered.push((peer, addr));
             }
@@ -353,7 +356,7 @@ where
         let mut expired = Vec::new();
         self.discovered_nodes.retain(|(peer, addr, expiration)| {
             if *expiration <= now {
-                log::info!("expired: {} {}", peer, addr);
+                tracing::info!(%peer, address=%addr, "expired peer on address");
                 expired.push((*peer, addr.clone()));
                 return false;
             }