+++ protocols/relay/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.17.0 - unreleased
+## 0.17.0
 
 - Don't close connections on protocol failures within the relay-server.
   To achieve this, error handling was restructured:
@@ -10,7 +10,7 @@
 - Fix a rare race condition when making a reservation on a relay that could lead to a failed reservation.
   See [PR 4747](https://github.com/libp2p/rust-lib2pp/pulls/4747).
 - Propagate errors of relay client to the listener / dialer.
-  A failed reservation will now appear as `SwarmEvent::ListenerClosed` with the `ListenerId` of the corresponding `Swarm::listen_on` call. 
+  A failed reservation will now appear as `SwarmEvent::ListenerClosed` with the `ListenerId` of the corresponding `Swarm::listen_on` call.
   A failed circuit request will now appear as `SwarmEvent::OutgoingConnectionError` with the `ConnectionId` of the corresponding `Swarm::dial` call.
   Lastly, a failed reservation or circuit request will **no longer** close the underlying relay connection.
   As a result, we remove the following enum variants:
@@ -18,7 +18,7 @@
   - `relay::client::Event::OutboundCircuitReqFailed`
   - `relay::client::Event::InboundCircuitReqDenied`
   - `relay::client::Event::InboundCircuitReqDenyFailed`
-  
+
   See [PR 4745](https://github.com/libp2p/rust-lib2pp/pulls/4745).
 
 ## 0.16.2